requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
cbor2==5.6.4
cffi==1.16.0
cryptography==42.0.8
Django==4.2.13
django-appconf==1.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
pycparser==2.22
pyOpenSSL==24.1.0
sqlparse==0.5.0
typing_extensions==4.12.2
webauthn==2.1.0
